Revert "Merge pull request #1 from Adidier/master"
This reverts commit 4edca6e0baff5f4d2179d279af4f1db41d4da5d5, reversing
changes made to 2cdc3dff50c0eec3b549ae5405e6854d88923498.
