--- ComputoGrafico/Assets/Shaders/Default/default.frag
@@ -1,10 +1,0 @@
-#version 330
-
-in vec4 vCol;
-
-out vec4 colour;
-
-void main()
-{
-	colour = vCol;
-}
