+++ ComputoGrafico/Assets/Shaders/Menu/shader.vert
@@ -6,13 +6,9 @@ out vec4 vCol;
 
 uniform mat4 model;
 uniform mat4 projection;
-uniform mat4 view;
-
-
-
 
 void main()
 {
-	gl_Position = projection * view *  model * vec4(pos, 1.0);
-	vCol = vec4(1.0f, 0.0f, 0.0f, 1.0f);
+	gl_Position = projection * model * vec4(pos, 1.0);
+	vCol = vec4(clamp(pos, 0.0f, 1.0f), 1.0f);
 }
