+++ ComputoGrafico/BCE/GameState.cpp
