+++ ComputoGrafico/BCE/GameState.h
@@ -1,16 +1,15 @@
 #pragma once
 #include "Platform.h"
-#include <map>
+
 class Platform;
 class GameStateManager;
 
 class GameState
 {
 public:
-	virtual void Init() = 0;
+	virtual void Init(Platform* platform, GameStateManager* manager) = 0;
 	virtual void Draw() = 0;
-	virtual bool Input(std::map<int, bool> keys) = 0;
-	virtual bool MouseInput(int x, int y, bool leftbutton) = 0;
+	virtual bool Input(int keyInput) = 0;
 	virtual void Update() = 0;
 	virtual void Close() = 0;
 };
