+++ ComputoGrafico/BCE/GameStateManager.h
@@ -10,11 +10,9 @@ class GameStateManager
 private:
 	std::stack<GameState*> states;
 	Platform* platform;
+public:
 	GameStateManager();
 	~GameStateManager();
-	static GameStateManager* ptr;
-public:
-	static GameStateManager* getPtr();
 	void GameLoop();
 	void SetState(GameState* state);
 	void RealaseState();
