+++ ComputoGrafico/BCE/Shader.cpp
@@ -79,7 +79,6 @@ void Shader::CompileShader(const char* vertexCode, const char* fragmentCode)
 
 	uniformProjection = glGetUniformLocation(shaderID, "projection");
 	uniformModel = glGetUniformLocation(shaderID, "model");
-	uniformView = glGetUniformLocation(shaderID, "view");
 }
 
 GLuint Shader::GetProjectionLocation()
@@ -89,10 +88,6 @@ GLuint Shader::GetProjectionLocation()
 GLuint Shader::GetModelLocation()
 {
 	return uniformModel;
-}
-GLuint Shader::GetViewLocation()
-{
-	return uniformView;
 }
 
 void Shader::UseShader()
