+++ ComputoGrafico/BCE/Shader.h
@@ -19,7 +19,6 @@ public:
 
 	GLuint GetProjectionLocation();
 	GLuint GetModelLocation();
-	GLuint GetViewLocation();
 
 	void UseShader();
 	void ClearShader();
@@ -27,7 +26,7 @@ public:
 	~Shader();
 
 private:
-	GLuint shaderID, uniformProjection, uniformModel, uniformView;
+	GLuint shaderID, uniformProjection, uniformModel;
 
 	void CompileShader(const char* vertexCode, const char* fragmentCode);
 	void AddShader(GLuint theProgram, const char* shaderCode, GLenum shaderType);
