+++ ComputoGrafico/ComputoGrafico.vcxproj.filters
@@ -5,95 +5,54 @@
       <UniqueIdentifier>{67DA6AB6-F800-4c08-8B7A-83BB121AAD01}</UniqueIdentifier>
       <Extensions>rc;ico;cur;bmp;dlg;rc2;rct;bin;rgs;gif;jpg;jpeg;jpe;resx;tiff;tif;png;wav;mfcribbon-ms</Extensions>
     </Filter>
+    <Filter Include="BCE">
+      <UniqueIdentifier>{0039e17c-9ad2-445f-9baf-755fc38d0879}</UniqueIdentifier>
+    </Filter>
     <Filter Include="Game">
       <UniqueIdentifier>{2317a8ee-ecfa-4357-ba84-532eb69df2b7}</UniqueIdentifier>
     </Filter>
-    <Filter Include="CGE">
-      <UniqueIdentifier>{0039e17c-9ad2-445f-9baf-755fc38d0879}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="CGE\Base">
-      <UniqueIdentifier>{b8081ddd-fda0-4507-bb0d-ff5b3bd13941}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="CGE\Graphic">
-      <UniqueIdentifier>{d8619dec-de55-4ad0-a680-87a61f70b8db}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Assets">
-      <UniqueIdentifier>{e982d58c-17a0-4021-a688-a70118ffcdb8}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Assets\Shaders">
-      <UniqueIdentifier>{98c012a2-0e8b-4e12-8b05-465a584724bd}</UniqueIdentifier>
-    </Filter>
-    <Filter Include="Assets\Shaders\Default">
-      <UniqueIdentifier>{1e3942ff-49fe-4a7d-8a74-46ba389ebe19}</UniqueIdentifier>
-    </Filter>
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="BCE\GameState.cpp">
+      <Filter>BCE</Filter>
+    </ClCompile>
+    <ClCompile Include="BCE\GameStateManager.cpp">
+      <Filter>BCE</Filter>
+    </ClCompile>
+    <ClCompile Include="BCE\Platform.cpp">
+      <Filter>BCE</Filter>
+    </ClCompile>
     <ClCompile Include="main.cpp">
       <Filter>Game</Filter>
     </ClCompile>
     <ClCompile Include="Menu.cpp">
       <Filter>Game</Filter>
     </ClCompile>
-    <ClCompile Include="CGE\Base\GameState.cpp">
-      <Filter>CGE\Base</Filter>
+    <ClCompile Include="BCE\Mesh.cpp">
+      <Filter>BCE</Filter>
     </ClCompile>
-    <ClCompile Include="CGE\Base\GameStateManager.cpp">
-      <Filter>CGE\Base</Filter>
-    </ClCompile>
-    <ClCompile Include="CGE\Base\Platform.cpp">
-      <Filter>CGE\Base</Filter>
-    </ClCompile>
-    <ClCompile Include="CGE\Base\ShaderManager.cpp">
-      <Filter>CGE\Base</Filter>
-    </ClCompile>
-    <ClCompile Include="CGE\Graphic\Camera.cpp">
-      <Filter>Game</Filter>
-    </ClCompile>
-    <ClCompile Include="CGE\Graphic\Mesh.cpp">
-      <Filter>CGE\Graphic</Filter>
-    </ClCompile>
-    <ClCompile Include="CGE\Graphic\Shader.cpp">
-      <Filter>CGE\Graphic</Filter>
+    <ClCompile Include="BCE\Shader.cpp">
+      <Filter>BCE</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="BCE\Platform.h">
+      <Filter>BCE</Filter>
+    </ClInclude>
+    <ClInclude Include="BCE\GameState.h">
+      <Filter>BCE</Filter>
+    </ClInclude>
+    <ClInclude Include="BCE\GameStateManager.h">
+      <Filter>BCE</Filter>
+    </ClInclude>
     <ClInclude Include="Menu.h">
       <Filter>Game</Filter>
     </ClInclude>
-    <ClInclude Include="CGE\Base\GameState.h">
-      <Filter>CGE\Base</Filter>
+    <ClInclude Include="BCE\Mesh.h">
+      <Filter>BCE</Filter>
     </ClInclude>
-    <ClInclude Include="CGE\Base\GameStateManager.h">
-      <Filter>CGE\Base</Filter>
+    <ClInclude Include="BCE\Shader.h">
+      <Filter>BCE</Filter>
     </ClInclude>
-    <ClInclude Include="CGE\Base\Platform.h">
-      <Filter>CGE\Base</Filter>
-    </ClInclude>
-    <ClInclude Include="CGE\Base\ShaderManager.h">
-      <Filter>CGE\Base</Filter>
-    </ClInclude>
-    <ClInclude Include="CGE\Graphic\Camera.h">
-      <Filter>Game</Filter>
-    </ClInclude>
-    <ClInclude Include="CGE\Graphic\Mesh.h">
-      <Filter>CGE\Graphic</Filter>
-    </ClInclude>
-    <ClInclude Include="CGE\Graphic\Shader.h">
-      <Filter>CGE\Graphic</Filter>
-    </ClInclude>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Assets\Shaders\Default\default.frag">
-      <Filter>Assets\Shaders\Default</Filter>
-    </None>
-    <None Include="Assets\Shaders\Default\default.vert">
-      <Filter>Assets\Shaders\Default</Filter>
-    </None>
-    <None Include="Assets\Shaders\OneColor.frag">
-      <Filter>Assets\Shaders</Filter>
-    </None>
-    <None Include="Assets\Shaders\OneColor.vert">
-      <Filter>Assets\Shaders</Filter>
-    </None>
   </ItemGroup>
 </Project>
