+++ ComputoGrafico/Menu.cpp
@@ -14,17 +14,13 @@ Menu::~Menu()
 {
 }
 
-void Menu::Init()
+void Menu::Init(Platform* platform, GameStateManager* manager)
 {
 	std::cout << " Menu Init" << std::endl;
-	this->platform = Platform::GetPtr();
-	this->manager = GameStateManager::getPtr();
-	camera = Camera(glm::vec3(0, 0, 0.6f), glm::vec3(0.0f, 1.0f, 0.0f), 0.0f, 0.0f, 55.0f, 0.1f);
-	shaderManager = ShaderManager::getPtr();
-	shaderManager->initShader(&camera);
-	shaderManager->LoadShaders("OneColor", "Assets/Shaders/OneColor.vert", "Assets/Shaders/OneColor.frag");
+	this->platform = platform;
+	this->manager = manager;
+	LoadShaders();
 	LoadModels();
-	
 }
 
 void Menu::LoadModels()
@@ -61,46 +57,49 @@ void Menu::LoadModels()
 }
 void Menu::LoadShaders()
 {
-	
+	Shader *shader1 = new Shader();
+	shader1->CreateFromFiles(vShader, fShader);
+	shaderList.push_back(*shader1);
 }
 
 void Menu::Draw()
 {
 	platform->RenderClear();
-	shaderManager->Activate("OneColor");
-	shaderManager->draw();
+	
+	GLuint uniformProjection = 0, uniformModel = 0;
+	glm::mat4 projection = glm::perspective(45.0f, (float)platform->GetWidth() / platform->GetHeight(), 0.1f, 100.0f);
+	shaderList[0].UseShader();
+	uniformModel = shaderList[0].GetModelLocation();
+	uniformProjection = shaderList[0].GetProjectionLocation();
+
 	glm::mat4 model(1);
-	GLint uniformModel = shaderManager->GetModelLocation();
+
 	model = glm::translate(model, glm::vec3(0.0f, 0.0f, -2.5f));
 	model = glm::scale(model, glm::vec3(0.4f, 0.4f, 1.0f));
 	glUniformMatrix4fv(uniformModel, 1, GL_FALSE, glm::value_ptr(model));
+	glUniformMatrix4fv(uniformProjection, 1, GL_FALSE, glm::value_ptr(projection));
 	meshList[0]->RenderMesh();
 
 
-	shaderManager->Activate("default");
-	shaderManager->draw();
-	uniformModel = shaderManager->GetModelLocation();
-	model = glm::translate(model, glm::vec3(10.0f, 0.0f, -8.5f));
-	model = glm::scale(model, glm::vec3(0.7f, 0.4f, 1.0f));
-	glUniformMatrix4fv(uniformModel, 1, GL_FALSE, glm::value_ptr(model));
-	meshList[0]->RenderMesh();
+	glUseProgram(0);
+
 
 	platform->RenderPresent();
 }
-bool Menu::MouseInput(int x, int y, bool leftbutton)
-{
-	if (x != -1 || y != -1)
-		camera.mouseControl(x, y);
 
+bool Menu::Input(int keyInput)
+{
+	std::cout << " Menu Input" << std::endl;
+	if (keyInput == 27)
+	{
+	//	exit(1);
+	}
+	else
+	{
+
+	}
 	return false;
 }
-
-bool Menu::Input(std::map<int, bool> keys)
-{
-	camera.keyControl(keys, platform->GetDeltaTime());
-	return false;
-}
-
 
 void Menu::Update()
 {
